b4860: Add support frame work for dsp loader
 * New frame work for dsp loader is added
 * All the existing dsp loader are integrated to it.

Reviewed-on: http://git.am.freescale.net:8181/2791
Tested-by: Burmi Naveen-B16502 <NaveenBurmi@freescale.com>

+++ ipc/include/fsl_ipc_um.h
@@ -27,7 +27,7 @@ typedef struct {
 	int		init;
 	/* MAX TXreq Linerized TB buffer size */
 	uint32_t 	max_txreq_lbuff_size;
-	uint32_t 	txreq_tb_lbuff_paddr;
+	unsigned long 	txreq_tb_lbuff_paddr;
 
 	int 		num_channels;
 	ipc_channel_us_t *channels[MAX_CHANNELS];
